chore(dashboard): rewrite html/css main structure
