+++ dashboard/src/components/Filters.js
@@ -187,6 +187,7 @@ const Container = styled.div`
   display: flex;
   justify-content: center;
   border-bottom: 1px solid #ddd;
+  width: 100%;
 `;
 
 const Subcontainer = styled.div`
