+++ dashboard/src/components/ScrollToTop.js
@@ -6,10 +6,10 @@ const ScrollToTop = () => {
 
   useEffect(() => {
     // https://stackoverflow.com/questions/51517324/scrollto-method-doesnt-work-in-edge
-    if (document?.querySelector('.main')?.scrollTo) {
-      document.querySelector('.main').scrollTo(0, 0);
-    } else if (document?.querySelector('.main').scrollTop) {
-      document.querySelector('.main').scrollTop = 0;
+    if (document?.querySelector('.main-content')?.scrollTo) {
+      document?.querySelector('.main-content')?.scrollTo(0, 0);
+    } else if (document?.querySelector('.main-content')?.scrollTop) {
+      document.querySelector('.main-content').scrollTop = 0;
     }
   }, [pathname]);
