+++ dashboard/src/index.scss
@@ -32,26 +32,50 @@ li {
   display: flex;
   flex-direction: column;
   height: 100vh;
+  overflow: hidden;
+  // border: 10px solid black;
 }
 
 .main {
-  flex: 1;
-  height: 100%;
-  flex-direction: column;
+  flex-grow: 1;
+  flex-basis: 100%;
   display: flex;
-  height: 100%;
-  overflow-y: auto;
+  overflow: hidden;
+  // border: 10px solid orange;
+  width: 100%;
 }
 
 .main-container {
-  display: flex;
+  flex: 1;
+  flex-basis: 100%;
   height: 100%;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  // border: 10px solid red;
+  align-items: center;
+  justify-content: center;
 }
 
 .main-content {
-  height: 100%;
-  padding: 0 68px;
+  flex-grow: 1;
+  flex-basis: 100%;
+  display: flex;
+  flex-direction: column;
+  overflow: auto;
+  padding: 0px 4rem 3rem;
+  // border: 10px solid black;
 }
+
+// .container {
+//   width: 100%;
+//   padding-left: 0 !important;
+//   padding-right: 0 !important;
+//   padding-bottom: 40px;
+//   position: relative;
+//   display: flex;
+//   flex-direction: column;
+// }
 
 input:focus,
 textarea:focus,
@@ -127,11 +151,17 @@ hr {
     height: auto;
   }
 
+  #root {
+    overflow-y: auto;
+    height: auto;
+  }
+
   .main-content {
     height: auto;
     margin-left: 0 !important;
     max-width: 100%;
     padding: 0 !important;
+    overflow: initial;
   }
   .report-container {
     padding: 0 !important;
