+++ dashboard/src/scenes/person/list.js
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useContext, useEffect, useState } from 'react';
-import { Col, Container, Row } from 'reactstrap';
+import { Col, Row } from 'reactstrap';
 import { useHistory } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
@@ -56,7 +56,7 @@ const List = () => {
   const total = personsFiltered.length;
 
   return (
-    <Container>
+    <>
       <Header
         title={
           <>
@@ -158,7 +158,7 @@ const List = () => {
         ]}
       />
       <Page page={page} limit={limit} total={total} onChange={({ page }) => setPage(page, true)} />
-    </Container>
+    </>
   );
 };
