+++ dashboard/src/scenes/territory/view.js
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React from 'react';
-import { Container, Row, Col, FormGroup, Input, Label } from 'reactstrap';
+import { Row, Col, FormGroup, Input, Label } from 'reactstrap';
 import { useParams, useHistory } from 'react-router-dom';
 import { toastr } from 'react-redux-toastr';
 import { Formik } from 'formik';
@@ -39,7 +39,7 @@ const View = () => {
   if (!territory) return <Loading />;
 
   return (
-    <Container style={{ padding: '40px 0' }}>
+    <>
       <Header
         title={<BackButton />}
         onRefresh={() =>
@@ -105,7 +105,7 @@ const View = () => {
         </Formik>
       </Box>
       <Observations territory={territory || { _id: id }} />
-    </Container>
+    </>
   );
 };
